Massive updates!
Pylinted all the things, some more. Refactored ringbond to ringlink. Added
in stuff for hydrogens, for charge, for isotopes, and in general for
dealing with bracketed atoms. Can now enter the entire periodic table,
specify wacky amounts of hydrogen for individual molecules, and specify
charge, all in a way that will still be represented. Whee!

In toMolecule, unfinished for now and likely to remain unfinished for some
time longer: Chirality. Cis/trans. Aromaticity.