Added main index; edits api/views.py
Made some tidying edits to api/views.py. Created JsonResponse and
SvgResponse helper methods, which return HttpResponse objects with the
right pretty-printing and content-type for when we're returning JSON or
SVG.

Created a new template in orgo/templates/orgo/index.html, to be the main
index of the site (which up until now didn't exist) -- with links to the
api, app, and orgo subdivisions. Useful mainly for testing purposes.

Created methods RequiredQueryParamsResponse and ModelNotFoundResponse for
returning some errors conveniently, too.